Re-design the data schema and import module

+++ src/main/java/edu/neu/cs6510/pnnl/hunting/model/Vav.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class Vav {
     private Integer id;
 
-    private Integer zoneAirFlow;
+    private Double zoneAirFlow;
 
     private Double zoneCoolingAirFlowSetPoint;
